+++ Source/3dStuff.cpp
@@ -11,7 +11,7 @@
 #include "GraphicLoader.h"
 #include "Log.h"
 #include "Script.h"
-#include "ScriptFunctions.h"
+#include "ScriptReservedFunctions.h"
 #include "Text.h"
 #include "Timer.h"