+++ Source/CritterManager.cpp
@@ -9,7 +9,7 @@
 #include "Log.h"
 #include "Random.h"
 #include "Script.h"
-#include "ScriptFunctions.h"
+#include "ScriptReservedFunctions.h"
 #include "Text.h"
 #include "Timer.h"
 #include "Utils.h"