+++ Source/Shared/Ini.cpp
@@ -74,7 +74,7 @@ inline void Cleanup( string& str )
 
 Ini::Ini() : KeepComments( false ), KeepSectionsRaw( false ), KeepKeysOrder( false )
 {
-    #if defined (FO_WINDOWS) && defined (FO_MSVC) && _MSC_VER >= 1910 && !defined (__INTELLISENSE__)
+    #if defined (FOCLASSIC_ENGINE) && defined (FO_WINDOWS) && defined (FO_MSVC) && _MSC_VER >= 1910 && !defined (__INTELLISENSE__)
     # pragma TODO("Use initializer list for CommentChars")
     #endif
     if( CommentChars.empty() )